Building up realm tests
