+++ .github/workflows/swift.yml
@@ -11,9 +11,7 @@ on:
 
 jobs:
   build:
-
     runs-on: macos-14
-
     steps:
     - uses: actions/checkout@v4
     - name: Select Xcode  
@@ -58,7 +56,7 @@ jobs:
         run: >
           ls -l ./
       - name: Run sonar-scanner
-        uses: SonarSource/sonarcloud-github-action@v2.0.2
+        uses: SonarSource/sonarcloud-github-action@master
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
