+++ Tests/SearchOpsTests/RequestTests/ElasticSearch.v6/MappingTests.v6.swift
@@ -13,8 +13,11 @@ import SwiftyJSON
 @available(iOS 16.0.0, *)
 final class ElasticSearch_v6_MappingTests: XCTestCase {
   
-  override func setUpWithError() throws {
-      try? RealmManager.DeleteRealmDatabase()
+  override func setUp() {
+    super.setUp()
+    // TODO
+    // Use in memory realm
+    // https://www.mongodb.com/docs/atlas/device-sdks/sdk/swift/test-and-debug/
   }
   
   @MainActor
@@ -28,6 +31,3 @@ final class ElasticSearch_v6_MappingTests: XCTestCase {
     
   }
 }
-
-
-
