+++ Tests/SearchOpsTests/RequestTests/ElasticSearch.v8/MappingTests.v8.swift
@@ -13,8 +13,11 @@ import SwiftyJSON
 @available(iOS 16.0.0, *)
 final class ElasticSearch_v8_MappingTests: XCTestCase {
   
-  override func setUpWithError() throws {
-      try? RealmManager.DeleteRealmDatabase()
+  override func setUp() {
+    super.setUp()
+    // TODO
+    // Use in memory realm
+    // https://www.mongodb.com/docs/atlas/device-sdks/sdk/swift/test-and-debug/
   }
   
   @MainActor
